requirements.txt
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
urllib3==2.3.0
